+++ NEWS.md
@@ -5,6 +5,8 @@
 - Added support for aggregating an assay inside of a LongTable class object
 - Some in-progress updates to the CoreSet constructor which will be completed for the Fall release
 - Fixed an error in treatmentNames example
+- Fixed roxygen2 documentation warnings about S4 method documentation
+- Overhauled LongTable coerce methods to use the LongTableDataMapper class instead of the deprecated 'LongTable.config' attribute
 
 ## v1.5.4
 - Fix bug in `$<-` and `[[<-` methods where value was returned instead of updated object